examples/task_to_generate_data_transmission_report/requirements.txt
jinja2
robotframework
robotframework-statemachinelibrary
